Add agent reprompt on auctions page
Since grounding and structured output can't work together, the format for the final price has to be set in the agent's system instructions. However this is brittle, so if an agent fails to follow the format it will be skipped.

A button to reprompt unsuccessful agents have been added in the auctions page